rename aimanager to explosivemanager
